refactor: annual actual data ids and filepaths upd
Updated the filepaths and catalog ids for annual actual data
