+++ conf/base/pipelines/data_engineering/parameters.yml
@@ -1,5 +1,1 @@
-# This is a boilerplate parameters config generated for pipeline 'data_engineering'
-# using Kedro 0.16.4.
-#
-# Documentation for this file format can be found in
-# https://kedro.readthedocs.io/en/stable/04_user_guide/03_configuration.html#parameters
+
